+++ R/sjTabOdds.R
@@ -229,6 +229,7 @@ sjt.glm <- function(...,
                     showAbbrHeadline = TRUE,
                     showPseudoR = FALSE,
                     showICC = FALSE,
+                    showREvar = FALSE,
                     showLogLik = FALSE,
                     showAIC = FALSE,
                     showAICc = FALSE,
@@ -309,8 +310,8 @@ sjt.glm <- function(...,
   css.lasttablerow <- "border-bottom: double;"
   css.topborder <- "border-top:double;"
   css.depvarhead <- "text-align:center; border-bottom:1px solid;"
-  css.topcontentborder <- "border-top:2px solid;"
-  css.annorow <- "border-top:2px solid;"
+  css.topcontentborder <- "border-top:1px solid;"
+  css.annorow <- "border-top:1px solid;"
   css.noannorow <- "border-bottom:double;"
   css.annostyle <- "text-align:right;"
   css.leftalign <- "text-align:left;"
@@ -393,13 +394,13 @@ sjt.glm <- function(...,
   # ------------------------
   lmerob <- any(class(input_list[[1]]) == "glmerMod")
   if (lmerob && !requireNamespace("lme4", quietly = TRUE)) {
-    stop("Package 'lme4' needed for this function to work. Please install it.", call. = FALSE)
+    stop("Package `lme4` needed for this function to work. Please install it.", call. = FALSE)
   }
   # ------------------------
   # should AICc be computed? Check for package
   # ------------------------
   if (showAICc && !requireNamespace("AICcmodavg", quietly = TRUE)) {
-    warning("Package 'AICcmodavg' needed to show AICc. Argument 'showAICc' will be ignored.", call. = FALSE)
+    warning("Package `AICcmodavg` needed to show AICc. Argument `showAICc` will be ignored.", call. = FALSE)
     showAICc <- FALSE
   }
   # ------------------------
@@ -900,16 +901,84 @@ sjt.glm <- function(...,
   if (lmerob) {
     # css attribute "topcontentborder" already in this table row
     page.content <- paste0(page.content, sprintf("  <tr>\n    <td colspan=\"%i\" class=\"tdata summary leftalign randomparts\">Random Parts</td>\n  </tr>\n", headerColSpan + 1))
-    # first models indicates grouping levels
-    # we have to assume comparable models with same
-    # random intercepts
-    # count all random intercepts of all models
+    # -------------------------------------
+    # we need to know max amount of groups
+    # -------------------------------------
     all_mm_counts <- unlist(lapply(input_list, function(x) length(lme4::getME(x, "flist"))))
     # retrieve maximum random intercepts
     mmcount <- max(all_mm_counts)
     # get random intercepts from model with most intercepts
     mmgrps <- lme4::getME(input_list[[which.max(all_mm_counts)]], "flist")
-    # iterate grouping levels
+    # -------------------------------------
+    # show variance components?
+    # -------------------------------------
+    if (isTRUE(showREvar)) {
+      # -------------------------------------
+      # lets check which mdoels have random slopes, needed later
+      # -------------------------------------
+      has_rnd_slope <- unlist(lapply(input_list, function(mo) {
+        lapply(lme4::VarCorr(mo), function(x) dim(attr(x, "correlation"))[1] > 1)
+      }))
+      # -------------------------
+      # between-group variance
+      # -------------------------
+      # first models indicates grouping levels. we have to assume comparable models 
+      # with same random intercepts.
+      for (gl in 1:mmcount) {
+        page.content <- paste0(page.content, 
+                               sprintf("\n  <tr>\n    <td class=\"tdata summary leftalign\">&tau;<sub>00, %s</sub></td>\n", 
+                                       names(mmgrps[gl])))
+        # iterate models
+        for (i in 1:length(input_list)) {
+          # -------------------------
+          # insert "separator column"
+          # -------------------------
+          page.content <- paste0(page.content, "<td class=\"separatorcol\">&nbsp;</td>")
+          # get random intercept variance
+          reva <- lme4::VarCorr(input_list[[i]])
+          vars <- lapply(reva, function(x) x[[1]])
+          tau.00 <- sapply(vars, function(x) x[1])
+          if (length(tau.00) >= gl) {
+            rand.int.var <- paste0(sprintf("%.*f", digits.summary, tau.00[gl], collapse = ""))
+            page.content <- paste0(page.content, colspanstring, rand.int.var, "</td>\n")
+            
+          } else {
+            page.content <- paste(page.content, sprintf("   %s&nbsp;</td>\n", colspanstring))
+          }            
+        }
+        page.content <- paste0(page.content, "  </tr>\n")
+      }
+      # -------------------------
+      # finally, random slope intercept correlation
+      # -------------------------
+      if (any(has_rnd_slope)) {
+        # iterate final models
+        page.content <- paste0(page.content, "\n  <tr>\n    <td class=\"tdata summary leftalign\">&rho;<sub>01</sub></td>\n")
+        # iterate models
+        for (i in 1:length(input_list)) {
+          # -------------------------
+          # insert "separator column"
+          # -------------------------
+          page.content <- paste0(page.content, "<td class=\"separatorcol\">&nbsp;</td>")
+          # does model have random slope?
+          if (has_rnd_slope[i]) {
+            # get slope-intercept correlation
+            reva <- lme4::VarCorr(input_list[[i]])
+            cor_ <- unlist(lapply(reva, function(x) attr(x, "correlation")[1, 2]))
+            rho.01 <- paste0(sprintf("%.*f", digits.summary, cor_[1], collapse = ""))
+            page.content <- paste0(page.content, colspanstring, rho.01, "</td>\n")
+          } else {
+            page.content <- paste(page.content, sprintf("   %s&nbsp;</td>\n", colspanstring))
+          }
+        }
+        page.content <- paste0(page.content, "  </tr>\n")
+      }
+    }
+    # -------------------------------------
+    # N of grouping levels
+    # -------------------------------------
+    # first models indicates grouping levels. we have to assume comparable models 
+    # with same random intercepts.
     for (gl in 1:mmcount) {
       page.content <- paste0(page.content, sprintf("\n  <tr>\n    <td class=\"tdata summary leftalign\">N<sub>%s</sub></td>", names(mmgrps[gl])))
       # iterate models
@@ -1274,7 +1343,12 @@ sjt.glm <- function(...,
 #' @note Computation of p-values (if necessary) are based on Wald chi-squared tests from the 
 #'         \code{Anova}-function of the \pkg{car}-package.
 #'         \cr \cr
-#'         Furthermore. see 'Notes' in \code{\link{sjt.frq}}.
+#'         The variance components of the random parts (see \code{showREvar}) are
+#'         denoted like:
+#'         \itemize{
+#'          \item between-group-variance: tau-zero-zero
+#'          \item random-slope-intercept-correlation: rho-zero-one
+#'          }
 #'  
 #' @details See 'Details' in \code{\link{sjt.frq}}.
 #'
@@ -1356,6 +1430,7 @@ sjt.glmer <- function(...,
                       showAbbrHeadline = TRUE,
                       showPseudoR = FALSE,
                       showICC = TRUE,
+                      showREvar = TRUE,
                       showLogLik = FALSE,
                       showAIC = FALSE,
                       showAICc = FALSE,
@@ -1384,7 +1459,7 @@ sjt.glmer <- function(...,
                  showConfInt = showConfInt, showStdError = showStdError, 
                  ci.hyphen = ci.hyphen, separateConfColumn = separateConfColumn, newLineConf = newLineConf, 
                  group.pred = group.pred, showAbbrHeadline = showAbbrHeadline, showPseudoR = showPseudoR, showICC = showICC, 
-                 showLogLik = showLogLik, showAIC = showAIC, showAICc = showAICc, showDeviance = showDeviance,
+                 showREvar = showREvar, showLogLik = showLogLik, showAIC = showAIC, showAICc = showAICc, showDeviance = showDeviance,
                  showChi2 = FALSE, showHosLem = showHosLem, showFamily = showFamily, remove.estimates = remove.estimates, 
                  cellSpacing = cellSpacing, cellGroupIndent = cellGroupIndent, encoding = encoding, 
                  CSS = CSS, useViewer = useViewer, no.output = no.output, remove.spaces = remove.spaces))